tests/wpt/web-platform-tests/orientation-event/t002-manual.https.html

<!DOCTYPE html>
<html>
  <head>
    <title>The corresponding event must be of type DeviceOrientationEvent and must fire on the window object.</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="author" title="Mosquito FP7">
    <link rel="reviewer author" title="Zhiqiang Zhang" href="mailto:zhiqiang.zhang@intel.com"> <!-- 2013-10-12 -->
  </head>
  <body>
    <p>Rotate the device to run the tests.</p>
    <div id="log"></div>
    <script>
      var t1 = async_test("The corresponding event must be of type DeviceOrientationEvent");
      var t2 = async_test("The corresponding event must fire on the window object");
      var run = false;
      window.addEventListener("deviceorientation", function(e) {
        if (!run) {
          run = true;
          var _this = this;
          t1.step(function() {
            assert_equals(e.type, "deviceorientation");
          });
          t1.done();
          t2.step(function() {
            assert_equals(_this, window);
          });
          t2.done();
        }
      }, false);
    </script>
  </body>
</html>
